refactor: 레이아웃, 페이지 분리
