+++ src/App.tsx
@@ -1,31 +1,19 @@
 import React, { useCallback, useState } from "react";
-import { generateItems } from "./utils";
-import {
-  Header,
-  ItemList,
-  ComplexForm,
-  NotificationSystem,
-} from "./components";
 import { User, Notification, Theme } from "./types/types";
 import { NotificationContext, ThemeContext, UserContext } from "./context";
 import { useMemo } from "./@lib";
+import { HomePage } from "./pages/HomePage";
+import { Layout } from "./components/Layout";
 
 // 메인 App 컴포넌트
 const App: React.FC = () => {
   const [theme, setTheme] = useState<Theme>("light");
-  const [items, setItems] = useState(() => generateItems(1000));
+
   const [user, setUser] = useState<User | null>(null);
   const [notifications, setNotifications] = useState<Notification[]>([]);
 
   const toggleTheme = useCallback(() => {
     setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
   }, []);
 
   const addNotification = useCallback(
@@ -77,22 +65,9 @@ const App: React.FC = () => {
     <ThemeContext.Provider value={themeValue}>
       <UserContext.Provider value={userValue}>
         <NotificationContext.Provider value={notificationValue}>
-          <div
-            className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-          >
-            <Header />
-            <div className="container mx-auto px-4 py-8">
-              <div className="flex flex-col md:flex-row">
-                <div className="w-full md:w-1/2 md:pr-4">
-                  <ItemList items={items} onAddItemsClick={addItems} />
-                </div>
-                <div className="w-full md:w-1/2 md:pl-4">
-                  <ComplexForm />
-                </div>
-              </div>
-            </div>
-            <NotificationSystem />
-          </div>
+          <Layout>
+            <HomePage />
+          </Layout>
         </NotificationContext.Provider>
       </UserContext.Provider>
     </ThemeContext.Provider>
